Use Hotwire for PokéPaste
