+++ app/views/trainees/index.haml
@@ -1,7 +1,5 @@
 - provide :title, "Trainees"
 
-= javascript_include_tag "pokepaste"
-
 #trainees-index{"data-controller": "trainees-index"}
   - if @trainees.blank?
     %p.grey-text Nothing to show here. Click the + below to get started!
@@ -33,27 +31,28 @@
       %h4 Confirm delete
       %p Are you sure? This will delete the selected trainees permanently.
       %label
-        = check_box_tag "confirm-delete-checkbox", nil, false, "data-action": "change->trainees-index#setConfirmDeleteState"
+        = check_box_tag "confirm-delete-checkbox", nil, false,
+                        "data-action": "change->trainees-index#setConfirmDeleteState"
         %span.black-text Yes, I understand and would like to delete these trainees.
     .modal-footer
       %a#cancel-delete.modal-close.waves-effect.waves-teal.btn-flat Cancel
       %a#confirm-delete.modal-close.waves-effect.waves-teal.btn-flat.disabled{"data-action": "click->trainees-index#delete"}
         Delete
 
-#add-pokepaste.modal
+#add-pokepaste.modal{"data-controller": "pokepaste"}
   .modal-content
     %h4 Add from PokéPaste
     .input-container
       .input-field.url
         = label_tag :url, "URL (optional)"
         = text_field_tag :url, nil, placeholder: "https://pokepast.es/",
-                         oninput: "updatePaste()"
+                         "data-action": "input->pokepaste#update"
       .input-field.paste
         %label PokéPaste
-        %textarea#paste.materialize-textarea{maxlength: 5000, oninput: "enablePasteButton()"}
+        %textarea#paste.materialize-textarea{maxlength: 5000, "data-action": "input->pokepaste#enableButton"}
       .footer-spacer
   .modal-footer
     %a#cancel-pokepaste.modal-close.waves-effect.waves-teal.btn-flat Cancel
-    %a#confirm-pokepaste.modal-close.waves-effect.waves-teal.btn-flat.disabled{onclick: "submitPaste()"} Add
+    %a#confirm-pokepaste.modal-close.waves-effect.waves-teal.btn-flat.disabled{"data-action": "click->pokepaste#submit"} Add
 
 = render "shared/ads"
